Patch POM to include compile dependencies
The ffmpeg decoder extension implicitly depends on the androidx.media3 media3-decoder and media3-exoplayer modules, but didn't publish this in its Maven POM. Because Gradle depends on this information for things like bytecode transformations, this could cause apps using this library to crash if required rewrites weren't applied to this library as well.
Thus, we now manually modify the POM to explicitly depend on the public versions of the library dependencies specified in androidx/media:libraries/decoder_ffmpeg/build.gradle.